Use DependencyService when needed instead
