+++ src/Controls/src/Core/IPlatformInvalidate.cs
@@ -1,7 +1,6 @@
 namespace Microsoft.Maui.Controls.Internals
 {
 	public interface IPlatformInvalidate
-
 	{
 		void Invalidate(VisualElement visualElement);
 	}
